refactor: 유효성 검사 관련 Usecase protocol 및 구현체 수정 #14
