--- iOS-Gaebalzarang/iOS-Gaebalzarang/Discriminator/ValidationSectionCase.swift
@@ -1,16 +1,0 @@
-//
-//  ValidationSectionCase.swift
-//  iOS-Gaebalzarang
-//
-//  Created by juntaek.oh on 2022/08/17.
-//
-
-import Foundation
-
-enum ValidationSectionCase {
-    
-    case idValidityCheck
-    case pswValidityCheck
-    case idOverrapValidation
-    case pswEqualValidation
-}
